Test file consolidation
